+++ motd.txt
@@ -1,5 +1,5 @@
 ************************************************************
-IPchanger last update 11.06.2020
+IPchanger last update 10.07.2020
 Geoip last update     06.12.2019
 Geoip based on        Dec 3 2019
 ************************************************************
