--- source-code-win/1.2.3/build.bat
@@ -1,5 +1,0 @@
-@echo off
-python setup-ipchanger.py py2exe
-move /Y dist\Lib\*.* dist\tcl
-rmdir dist\Lib
-ren dist\tcl Lib
