--- source-code-win/1.2.3/requirements/pip-requirements.txt
@@ -1,3 +1,0 @@
-urllib3
-pysocks
-py2exe
