+++ source-code-win/1.2.4/build.bat
@@ -1,0 +1,5 @@
+@echo off
+python setup-ipchanger.py py2exe
+move /Y dist\Lib\*.* dist\tcl
+rmdir dist\Lib
+ren dist\tcl Lib
