+++ source-code-win/1.2.4/requirements/pip-requirements.txt
@@ -1,0 +1,3 @@
+urllib3
+pysocks
+py2exe
