+++ version.txt
@@ -1,1 +1,1 @@
-1.2.3
+1.2.4
